+++ mmcv/utils/__init__.py
@@ -2,10 +2,11 @@
 # Copyright (c) Open-MMLab. All rights reserved.
 from .config import Config, ConfigDict, DictAction
 from .misc import (check_prerequisites, concat_list, deprecated_api_warning,
-                   import_modules_from_strings, is_list_of, is_seq_of, is_str,
-                   is_tuple_of, iter_cast, list_cast, requires_executable,
-                   requires_package, slice_list, to_1tuple, to_2tuple,
-                   to_3tuple, to_4tuple, to_ntuple, tuple_cast)
+                   import_modules_from_strings, is_list_of,
+                   is_method_overriden, is_seq_of, is_str, is_tuple_of,
+                   iter_cast, list_cast, requires_executable, requires_package,
+                   slice_list, to_1tuple, to_2tuple, to_3tuple, to_4tuple,
+                   to_ntuple, tuple_cast)
 from .path import (check_file_exist, fopen, is_filepath, mkdir_or_exist,
                    scandir, symlink)
 from .progressbar import (ProgressBar, track_iter_progress,
@@ -31,7 +32,8 @@ except ImportError:
         'digit_version', 'get_git_hash', 'import_modules_from_strings',
         'assert_dict_contains_subset', 'assert_attrs_equal',
         'assert_dict_has_keys', 'assert_keys_equal', 'check_python_script',
-        'to_1tuple', 'to_2tuple', 'to_3tuple', 'to_4tuple', 'to_ntuple'
+        'to_1tuple', 'to_2tuple', 'to_3tuple', 'to_4tuple', 'to_ntuple',
+        'is_method_overriden'
     ]
 else:
     from .env import collect_env
@@ -60,5 +62,5 @@ else:
         'get_git_hash', 'import_modules_from_strings', 'jit', 'skip_no_elena',
         'assert_dict_contains_subset', 'assert_attrs_equal',
         'assert_dict_has_keys', 'assert_keys_equal', 'assert_is_norm_layer',
-        'assert_params_all_zeros', 'check_python_script'
+        'assert_params_all_zeros', 'check_python_script', 'is_method_overriden'
     ]
